requirements.txt
PyMySQL==1.0.2
python-dotenv==1.0.0
